Change occurence of uint8_t to uint16_t to fix bug

+++ matrix.c
@@ -1,6 +1,6 @@
 #include "matrix.h"
 
-uint8_t find_vector_weight (vector_t x, uint8_t len) {
+uint8_t find_vector_weight (vector_t x, uint16_t len) {
     uint8_t weight = 0;
     for (int j = 0; j < len/8; j++) {
         for (int i = 0; i < 8; i++) {
